+++ backend/packages/files/parsers/code_python.py
@@ -4,11 +4,9 @@ from models import File
 from .common import process_file
 
 
-async def process_python(file: File, enable_summarization, brain_id, user_openai_api_key):
+async def process_python(file: File, brain_id):
     return await process_file(
         file=file,
         loader_class=PythonLoader,
-        enable_summarization=enable_summarization,
         brain_id=brain_id,
-        user_openai_api_key=user_openai_api_key,
     )