+++ backend/routes/chat/brainful_chat.py
@@ -2,7 +2,6 @@ from fastapi import HTTPException
 from llm.api_brain_qa import APIBrainQA
 from llm.qa_base import QABaseBrainPicking
 from models.brain_entity import BrainType
-from repository.brain import get_brain_details
 from repository.brain.get_brain_by_id import get_brain_by_id
 from routes.authorizations.brain_authorization import validate_brain_authorization
 from routes.authorizations.types import RoleEnum
@@ -27,11 +26,6 @@ class BrainfulChat(ChatInterface):
                 required_roles=[RoleEnum.Viewer, RoleEnum.Editor, RoleEnum.Owner],
             )
 
-    def get_openai_api_key(self, brain_id, user_id):
-        brain_details = get_brain_details(brain_id)
-        if brain_details:
-            return brain_details.openai_api_key
-
     def get_answer_generator(
         self,
         brain_id,
@@ -39,7 +33,6 @@ class BrainfulChat(ChatInterface):
         model,
         max_tokens,
         temperature,
-        user_openai_api_key,
         streaming,
         prompt_id,
         user_id,
@@ -59,7 +52,6 @@ class BrainfulChat(ChatInterface):
                 max_tokens=max_tokens,
                 temperature=temperature,
                 brain_id=brain_id,
-                user_openai_api_key=user_openai_api_key,
                 streaming=streaming,
                 prompt_id=prompt_id,
             )
@@ -70,7 +62,6 @@ class BrainfulChat(ChatInterface):
             max_tokens=max_tokens,
             temperature=temperature,
             brain_id=brain_id,
-            user_openai_api_key=user_openai_api_key,
             streaming=streaming,
             prompt_id=prompt_id,
             user_id=user_id,