+++ backend/routes/chat/interface.py
@@ -7,10 +7,6 @@ class ChatInterface(ABC):
         pass
 
     @abstractmethod
-    def get_openai_api_key(self, brain_id, user_id):
-        pass
-
-    @abstractmethod
     def get_answer_generator(
         self,
         brain_id,
@@ -18,7 +14,6 @@ class ChatInterface(ABC):
         model,
         max_tokens,
         temperature,
-        user_openai_api_key,
         streaming,
         prompt_id,
         user_id,