+++ frontend/app/chat/[chatId]/components/ActionsBar/components/ChatInput/components/ConfigModal/hooks/useConfigModal.ts
@@ -8,9 +8,13 @@ import {
   getChatsConfigFromLocalStorage,
   saveChatsConfigInLocalStorage,
 } from "@/lib/api/chat/chat.local";
-import { USER_DATA_KEY, USER_IDENTITY_DATA_KEY } from "@/lib/api/user/config";
+import { USER_DATA_KEY } from "@/lib/api/user/config";
 import { useUserApi } from "@/lib/api/user/useUserApi";
-import { defaultBrainConfig } from "@/lib/config/defaultBrainConfig";
+import {
+  defaultMaxTokens,
+  defaultModel,
+  defaultTemperature,
+} from "@/lib/config/defaultBrainConfig";
 import { useBrainContext } from "@/lib/context/BrainProvider/hooks/useBrainContext";
 import { ChatConfig } from "@/lib/context/ChatProvider/types";
 import { getAccessibleModels } from "@/lib/helpers/getAccessibleModels";
@@ -22,22 +26,18 @@ export const useConfigModal = () => {
   const [isConfigModalOpen, setIsConfigModalOpen] = useState(false);
   const { getBrain } = useBrainApi();
   const { currentBrainId } = useBrainContext();
-  const { getUser, getUserIdentity } = useUserApi();
+  const { getUser } = useUserApi();
 
   const { data: userData } = useQuery({
     queryKey: [USER_DATA_KEY],
     queryFn: getUser,
   });
-  const { data: userIdentityData } = useQuery({
-    queryKey: [USER_IDENTITY_DATA_KEY],
-    queryFn: getUserIdentity,
-  });
 
   const { register, watch, setValue } = useForm<ChatConfig>({
     defaultValues: {
-      model: defaultBrainConfig.model,
-      temperature: defaultBrainConfig.temperature,
-      maxTokens: defaultBrainConfig.maxTokens,
+      model: "gpt-3.5-turbo-16k",
+      temperature: 0,
+      maxTokens: 3000,
     },
   });
 
@@ -46,7 +46,6 @@ export const useConfigModal = () => {
   const maxTokens = watch("maxTokens");
 
   const accessibleModels = getAccessibleModels({
-    openAiKey: userIdentityData?.openai_api_key,
     userData,
   });
 
@@ -65,15 +64,12 @@ export const useConfigModal = () => {
       if (relatedBrainConfig === undefined) {
         return;
       }
-      setValue("model", relatedBrainConfig.model ?? defaultBrainConfig.model);
+      setValue("model", relatedBrainConfig.model ?? defaultModel);
       setValue(
         "temperature",
-        relatedBrainConfig.temperature ?? defaultBrainConfig.temperature
+        relatedBrainConfig.temperature ?? defaultTemperature
       );
-      setValue(
-        "maxTokens",
-        relatedBrainConfig.max_tokens ?? defaultBrainConfig.maxTokens
-      );
+      setValue("maxTokens", relatedBrainConfig.max_tokens ?? defaultMaxTokens);
     }
   }, []);